Show more detailed error messages (#27)
When sending requests from a front end app to the api, if a required attribute for the model you’re attempting to create with hangar is missing, the error message displays:

```
<NameError: undefined local variable or method 'env' for #<Hangar::ResourcesController:0x00007f7ee8b039c0>>
```

Removing this StandardError rescue surfaces the true error to us in the console. It will report back which fields are missing.

Another example: If I am trying to create a hospital model using hangar and the test environment migrations haven’t been run, the rails logs will show:

```
#<ActiveRecord::StatementInvalid: PG::UndefinedTable: ERROR:  relation "hospitals" does not exist
LINE 8:                WHERE a.attrelid = '"hospitals"'::regclass
                                          ^
:               SELECT a.attname, format_type(a.atttypid, a.atttypmod),
                     pg_get_expr(d.adbin, d.adrelid), a.attnotnull, a.atttypid, a.atttypmod,
                     c.collname, col_description(a.attrelid, a.attnum) AS comment
                FROM pg_attribute a
                LEFT JOIN pg_attrdef d ON a.attrelid = d.adrelid AND a.attnum = d.adnum
                LEFT JOIN pg_type t ON a.atttypid = t.oid
                LEFT JOIN pg_collation c ON a.attcollation = c.oid AND a.attcollation <> t.typcollation
               WHERE a.attrelid = '"hospitals"'::regclass
                 AND a.attnum > 0 AND NOT a.attisdropped
               ORDER BY a.attnum
>
```

In lieu of what had been showing:

```
<NameError: undefined local variable or method 'env' for #<Hangar::ResourcesController:0x00007f7ee8b039c0>>
```

Which is less indicative of what the actual problem is.